+++ behave4cmd0/log/steps.py
@@ -76,6 +76,7 @@ def make_log_record(category, level, message):
     logger = logging.getLogger(category)
     logger.log(level, message)
 
+
 def make_log_record_output(category, level, message,
                            format=None, datefmt=None, **kwargs):
     """
@@ -94,6 +95,7 @@ def make_log_record_output(category, level, message,
     record = logging.makeLogRecord(record_data)
     formatter = logging.Formatter(format, datefmt=datefmt)
     return formatter.format(record)
+
 
 class LogRecordTable(object):
 
@@ -132,6 +134,7 @@ class LogRecordTable(object):
 #         logger.log(current_level, "__LOG_RECORD__")
 
 @step('I create log records with')
+@step('I create log records with:')
 def step_I_create_logrecords_with_table(context):
     """
     Step definition that creates one more log records by using a table.
@@ -172,6 +175,7 @@ def step_I_create_logrecords_with_table(context):
 
 
 @step('I create a log record with')
+@step('I create a log record with:')
 def step_I_create_logrecord_with_table(context):
     """
     Create an log record by using a table to provide the parts.
@@ -182,7 +186,9 @@ def step_I_create_logrecord_with_table(context):
     assert len(context.table.rows) == 1, "REQUIRE: table.row.size == 1"
     step_I_create_logrecords_with_table(context)
 
+
 @step('I define the log record schema')
+@step('I define the log record schema:')
 def step_I_define_logrecord_schema_with_table(context):
     assert context.table, "REQUIRE: context.table"
     context.table.require_columns(["category", "level", "message"])
@@ -196,6 +202,7 @@ def step_I_define_logrecord_schema_with_table(context):
 
 
 @then('the command output should contain the following log records')
+@then('the command output should contain the following log records:')
 def step_command_output_should_contain_log_records(context):
     """
     Verifies that the command output contains the specified log records
@@ -221,6 +228,7 @@ def step_command_output_should_contain_log_records(context):
 
 
 @then('the command output should not contain the following log records')
+@then('the command output should not contain the following log records:')
 def step_command_output_should_not_contain_log_records(context):
     """
     Verifies that the command output contains the specified log records
@@ -244,7 +252,9 @@ def step_command_output_should_not_contain_log_records(context):
                 """
             '''.format(expected_output=output))
 
+
 @then('the command output should contain the following log record')
+@then('the command output should contain the following log record:')
 def step_command_output_should_contain_log_record(context):
     assert context.table, "REQUIRE: context.table"
     assert len(context.table.rows) == 1, "REQUIRE: table.row.size == 1"
@@ -252,12 +262,15 @@ def step_command_output_should_contain_log_record(context):
 
 
 @then('the command output should not contain the following log record')
+@then('the command output should not contain the following log record:')
 def step_command_output_should_not_contain_log_record(context):
     assert context.table, "REQUIRE: context.table"
     assert len(context.table.rows) == 1, "REQUIRE: table.row.size == 1"
     step_command_output_should_not_contain_log_records(context)
 
+
 @then('the command output should contain log records from categories')
+@then('the command output should contain log records from categories:')
 def step_command_output_should_contain_log_records_from_categories(context):
     """
     Verifies that the command output contains the specified log records
@@ -281,6 +294,7 @@ def step_command_output_should_contain_log_records_from_categories(context):
 
 
 @then('the command output should not contain log records from categories')
+@then('the command output should not contain log records from categories:')
 def step_command_output_should_not_contain_log_records_from_categories(context):
     """
     Verifies that the command output contains not log records from
@@ -302,7 +316,9 @@ def step_command_output_should_not_contain_log_records_from_categories(context):
     step_command_output_should_not_contain_log_records(context)
     context.table.remove_columns(["level", "message"])
 
+
 @then('the file "{filename}" should contain the log records')
+@then('the file "{filename}" should contain the log records:')
 def step_file_should_contain_log_records(context, filename):
     """
     Verifies that the command output contains the specified log records
@@ -322,7 +338,9 @@ def step_file_should_contain_log_records(context, filename):
         context.text = output
         step_file_should_contain_multiline_text(context, filename)
 
+
 @then('the file "{filename}" should not contain the log records')
+@then('the file "{filename}" should not contain the log records:')
 def step_file_should_not_contain_log_records(context, filename):
     """
     Verifies that the command output contains the specified log records
@@ -347,7 +365,9 @@ def step_file_should_not_contain_log_records(context, filename):
 def step_use_log_record_format_text(context, log_record_format):
     context.log_record_format = log_record_format
 
+
 @step('I use the log record configuration')
+@step('I use the log record configuration:')
 def step_use_log_record_configuration(context):
     """
     Define log record configuration parameters.