+++ features/steps/behave_active_tags_steps.py
@@ -32,6 +32,7 @@ def normalize_tags(tags):
 # STEP DEFINITIONS:
 # -----------------------------------------------------------------------------
 @given(u'I setup the current values for active tags with')
+@given(u'I setup the current values for active tags with:')
 def step_given_setup_the_current_values_for_active_tags_with(context):
     assert context.table, "REQUIRE: table"
     context.table.require_columns(["category", "value"])
@@ -48,6 +49,7 @@ def step_given_setup_the_current_values_for_active_tags_with(context):
 
 
 @then(u'the following active tag combinations are enabled')
+@then(u'the following active tag combinations are enabled:')
 def step_then_following_active_tags_combinations_are_enabled(context):
     assert context.table, "REQUIRE: table"
     assert context.active_value_provider, "REQUIRE: active_value_provider"
@@ -80,6 +82,7 @@ def step_then_following_active_tags_combinations_are_enabled(context):
 @step(u'unknown categories are ignored in active tags')
 def step_unknown_categories_are_ignored_in_active_tags(context):
     context.active_tags_ignore_unknown_categories = True
+
 
 @step(u'unknown categories are not ignored in active tags')
 def step_unknown_categories_are_not_ignored_in_active_tags(context):