+++ features/steps/behave_select_files_steps.py
@@ -49,38 +49,46 @@ class BasicBehaveRunner(object):
                 selected.append(six.text_type(filename))
         return selected
 
+
 # -----------------------------------------------------------------------------
 # STEP DEFINITIONS:
 # -----------------------------------------------------------------------------
 # pylint: disable=invalid-name
 @given('behave has the following feature fileset')
+@given('behave has the following feature fileset:')
 def step_given_behave_has_feature_fileset(context):
     assert context.text is not None, "REQUIRE: text"
     behave_runner = BasicBehaveRunner(config=copy(context.config))
     behave_runner.feature_files = FeatureListParser.parse(context.text)
     context.behave_runner = behave_runner
 
+
 @when('behave includes all feature files')
 def step_when_behave_includes_all_feature_files(context):
     assert context.behave_runner, "REQUIRE: context.behave_runner"
     context.behave_runner.config.include_re = None
+
 
 @when('behave includes feature files with "{pattern}"')
 def step_when_behave_includes_feature_files_with_pattern(context, pattern):
     assert context.behave_runner, "REQUIRE: context.behave_runner"
     context.behave_runner.config.include_re = re.compile(pattern)
 
+
 @when('behave excludes no feature files')
 def step_when_behave_excludes_no_feature_files(context):
     assert context.behave_runner, "REQUIRE: context.behave_runner"
     context.behave_runner.config.exclude_re = None
+
 
 @when('behave excludes feature files with "{pattern}"')
 def step_when_behave_excludes_feature_files_with_pattern(context, pattern):
     assert context.behave_runner, "REQUIRE: context.behave_runner"
     context.behave_runner.config.exclude_re = re.compile(pattern)
 
+
 @then('the following feature files are selected')
+@then('the following feature files are selected:')
 def step_then_feature_files_are_selected_with_text(context):
     assert context.text is not None, "REQUIRE: text"
     assert context.behave_runner, "REQUIRE: context.behave_runner"