+++ features/steps/behave_tag_expression_steps.py
@@ -37,6 +37,7 @@ class ModelElement(object):
         self.name = name
         self.tags = tags or []
 
+
 # -----------------------------------------------------------------------------
 # TYPE CONVERTERS:
 # -----------------------------------------------------------------------------
@@ -69,6 +70,7 @@ def step_given_the_tag_expression(context, tag_expression):
     """
     context.tag_expression = tag_expression
 
+
 @given('the default tags "{default_tags:TagExpression}"')
 def step_given_the_tag_expression(context, default_tags):
     """
@@ -83,7 +85,9 @@ def step_given_the_tag_expression(context, default_tags):
     if tag_expression is None:
         context.tag_expression = default_tags
 
+
 @then('the tag expression selects elements with tags')
+@then('the tag expression selects elements with tags:')
 def step_then_tag_expression_selects_elements_with_tags(context):
     """
     Checks if a tag expression selects an element with the given tags.
@@ -111,6 +115,7 @@ def step_then_tag_expression_selects_elements_with_tags(context):
 
 
 @given('the model elements with name and tags')
+@given('the model elements with name and tags:')
 def step_given_named_model_elements_with_tags(context):
     """
     .. code-block:: gherkin
@@ -141,6 +146,7 @@ def step_given_named_model_elements_with_tags(context):
 
 
 @then('the tag expression selects model elements with')
+@then('the tag expression selects model elements with:')
 def step_given_named_model_elements_with_tags(context):
     """
     .. code-block:: gherkin