+++ features/steps/i18n_emoji_steps.py
@@ -1,7 +1,9 @@
 # -*- coding: UTF-8 -*-
 # NEEDED-BY: features/i18n_emoji.feature
 
+from __future__ import absolute_import, print_function
 from behave import given
+
 
 @given(u'🎸')
 def step_impl(context):