+++ issue.features/steps/ansi_steps.py
@@ -10,9 +10,11 @@ from behave4cmd0.command_steps import \
 # ANSI CONTROL SEQUENCE INTRODUCER (CSI).
 CSI = u"\x1b["
 
+
 @then(u'the command output should contain ANSI escape sequences')
 def step_command_ouput_should_contain_ansi_sequences(context):
     step_command_output_should_contain_text(context, CSI)
+
 
 @then(u'the command output should not contain any ANSI escape sequences')
 def step_command_ouput_should_not_contain_ansi_sequences(context):