+++ issue.features/steps/behave_hooks_steps.py
@@ -3,6 +3,7 @@
 from __future__ import absolute_import
 from behave import then
 
+
 @then('the behave hook "{hook}" was called')
 def step_behave_hook_was_called(context, hook):
     substeps = u'Then the command output should contain "hooks.{0}: "'.format(hook)