+++ issue.features/steps/issue0766_steps.py
@@ -3,10 +3,13 @@
 from __future__ import print_function
 from behave import given
 
+
 @given(u'a step with table data')
+@given(u'a step with table data:')
 def step_with_table_data(ctx):
     assert ctx.table is not None, "REQUIRE: step.table"
 
+
 @given(u'a step with name="{name}"')
-def step_with_table_data(ctx, name):
+def step_with_name_param(ctx, name):
     print(u"name: {}".format(name))