+++ more.features/steps/tutorial_steps.py
@@ -1,16 +1,23 @@
-# -*- coding: utf-8 -*-
-"""Step implementations for tutorial example."""
+# -*- coding: UTF-8 -*-
+"""
+Step implementations for tutorial example.
+"""
 
-from behave import *
+from __future__ import absolute_import, print_function
+from behave import given, when, then
+from assertpy import assert_that
 
-@given('we have behave installed')
+
+@given(u'we have behave installed')
 def step_impl(context):
     pass
 
-@when('we implement a test')
-def step_impl(context):
-    assert True is not False
 
-@then('behave will test it for us!')
+@when(u'we implement a test')
 def step_impl(context):
-    assert context.failed is False
+    assert_that(True).is_not_equal_to(False)
+
+
+@then(u'behave will test it for us!')
+def step_impl(context):
+    assert_that(context.failed).is_equal_to(False)